requirements.txt
Flask
Flask-CORS
pdf2image
pdf2docx
docx
pandas
python-pptx
Pillow
